refactor general.py, split, repeat and tile in manipulation - extras (#582)
