+++ ivy/functional/backends/mxnet/old/general.py
@@ -20,7 +20,7 @@ from ivy.functional.ivy.device import default_device
 from ivy.functional.backends.mxnet.device import _callable_dev
 from ivy.functional.backends.mxnet.general import unstack
 from ivy.functional.backends.mxnet import _handle_flat_arrays_in_out, _mxnet_init_context,\
-    _scalar_or_flat_array_to_scalar, _handle_flat_arrays_in
+    _scalar_or_flat_array_to_scalar, _handle_flat_arrays_in, _flat_array_to_1_dim_array, _1_dim_array_to_flat_array
 
 #temporary imports
 from ivy.functional.backends.mxnet.general import linspace
@@ -142,44 +142,6 @@ def stack(xs, axis=0):
 
 
 
-
-
-def split(x, num_or_size_splits=None, axis=0, with_remainder=False):
-    if x.shape == ():
-        if num_or_size_splits is not None and num_or_size_splits != 1:
-            raise Exception('input array had no shape, but num_sections specified was {}'.format(num_or_size_splits))
-        return [x]
-    if num_or_size_splits == 1:
-        return [x]
-    elif with_remainder and isinstance(num_or_size_splits, int):
-        num_or_size_splits = x.shape[axis] if not num_or_size_splits else num_or_size_splits
-        num_chunks = x.shape[axis] / num_or_size_splits
-        num_chunks_int = _math.floor(num_chunks)
-        remainder_size = int((num_chunks - num_chunks_int) * num_or_size_splits)
-        num_or_size_splits = [num_or_size_splits]*num_chunks_int + [remainder_size]
-    if isinstance(num_or_size_splits, (list, tuple)):
-        csum = [0] + _np.cumsum(num_or_size_splits).tolist()
-        starts = csum[:-1]
-        ends = csum[1:]
-        if axis < 0:
-            slices = [tuple([Ellipsis, slice(s, e, 1)] + [slice(None, None, None)]*int(abs(axis)-1))
-                      for s, e in zip(starts, ends)]
-        else:
-            slices = [tuple([slice(None, None, None)]*axis + [slice(s, e, 1)])
-                      for s, e in zip(starts, ends)]
-        return [x[so] for so in slices]
-    return _mx.nd.split(x, x.shape[axis] if not num_or_size_splits else num_or_size_splits, axis)
-
-
-@_handle_flat_arrays_in_out
-def repeat(x, repeats, axis=None):
-    return _mx.nd.repeat(x, repeats, axis)
-
-
-def tile(x, reps):
-    if isinstance(reps, _mx.nd.ndarray.NDArray):
-        reps = reps.asnumpy().tolist()
-    return _mx.nd.tile(_flat_array_to_1_dim_array(x), reps)
 
 
 @_handle_flat_arrays_in
