+++ ivy/functional/backends/numpy/old/general.py
@@ -143,26 +143,6 @@ stack = _np.stack
 
 
 
-def split(x, num_or_size_splits=None, axis=0, with_remainder=False):
-    if x.shape == ():
-        if num_or_size_splits is not None and num_or_size_splits != 1:
-            raise Exception('input array had no shape, but num_sections specified was {}'.format(num_or_size_splits))
-        return [x]
-    if num_or_size_splits is None:
-        num_or_size_splits = x.shape[axis]
-    elif isinstance(num_or_size_splits, int) and with_remainder:
-        num_chunks = x.shape[axis] / num_or_size_splits
-        num_chunks_int = _math.floor(num_chunks)
-        remainder = num_chunks - num_chunks_int
-        if remainder != 0:
-            num_or_size_splits = [num_or_size_splits]*num_chunks_int + [int(remainder*num_or_size_splits)]
-    if isinstance(num_or_size_splits, (list, tuple)):
-        num_or_size_splits = _np.cumsum(num_or_size_splits[:-1])
-    return _np.split(x, num_or_size_splits, axis)
-
-
-repeat = _np.repeat
-tile = _np.tile
 constant_pad = lambda x, pad_width, value=0: _np.pad(_flat_array_to_1_dim_array(x), pad_width, constant_values=value)
 zero_pad = lambda x, pad_width: _np.pad(_flat_array_to_1_dim_array(x), pad_width)
 swapaxes = _np.swapaxes
