+++ ivy/functional/backends/tensorflow/old/general.py
@@ -115,35 +115,6 @@ def concatenate(xs, axis=-1):
 stack = _tf.stack
 
 
-def split(x, num_or_size_splits=None, axis=0, with_remainder=False):
-    if x.shape == ():
-        if num_or_size_splits is not None and num_or_size_splits != 1:
-            raise Exception('input array had no shape, but num_sections specified was {}'.format(num_or_size_splits))
-        return [x]
-    if num_or_size_splits is None:
-        dim_size = _tf.shape(x)[axis]
-        num_or_size_splits = dim_size
-    elif isinstance(num_or_size_splits, int) and with_remainder:
-        num_chunks = x.shape[axis] / num_or_size_splits
-        num_chunks_int = _math.floor(num_chunks)
-        remainder = num_chunks - num_chunks_int
-        if remainder != 0:
-            num_or_size_splits = [num_or_size_splits]*num_chunks_int + [int(remainder*num_or_size_splits)]
-    return _tf.split(x, num_or_size_splits, axis)
-
-
-repeat = _tf.repeat
-
-
-def tile(x, reps):
-    if x.shape == ():
-        x = _tf.reshape(x, (-1,))
-    if isinstance(reps, Number):
-        reps = [reps]
-    if isinstance(reps, Tensor) and reps.shape == ():
-        reps = _tf.reshape(reps, (-1,))
-    return _tf.tile(x, reps)
-
 
 def constant_pad(x, pad_width, value=0):
     if x.shape == ():
