+++ ivy/functional/backends/torch/old/general.py
@@ -104,42 +104,7 @@ def stack(xs: List[_torch.Tensor], axis: int = 0):
 
 
 
-def split(x, num_or_size_splits: Optional[Union[int, List[int]]] = None, axis: int = 0, with_remainder: bool = False)\
-        -> List[_torch.Tensor]:
-    if x.shape == ():
-        if num_or_size_splits is not None and num_or_size_splits != 1:
-            raise Exception('input array had no shape, but num_sections specified was {}'.format(num_or_size_splits))
-        return [x]
-    dim_size: int = x.shape[axis]
-    if num_or_size_splits is None:
-        # noinspection PyUnboundLocalVariable
-        num_or_size_splits = 1
-    elif isinstance(num_or_size_splits, int):
-        if with_remainder:
-            num_chunks = x.shape[axis] / num_or_size_splits
-            num_chunks_int = _math.floor(num_chunks)
-            remainder = num_chunks - num_chunks_int
-            if remainder == 0:
-                num_or_size_splits = _torch.round(_torch.tensor(dim_size) / _torch.tensor(num_or_size_splits))
-            else:
-                num_or_size_splits = tuple([num_or_size_splits] * num_chunks_int + [int(remainder*num_or_size_splits)])
-        else:
-            num_or_size_splits = _torch.round(_torch.tensor(dim_size) / _torch.tensor(num_or_size_splits))
-    elif isinstance(num_or_size_splits, list):
-        num_or_size_splits = tuple(num_or_size_splits)
-    return list(_torch.split(x, num_or_size_splits, axis))
 
-
-def repeat(x, repeats: Union[int, List[int]], axis: int = None):
-    if len(x.shape) == 0 and axis in [0, -1]:
-        axis = None
-    return _torch.repeat_interleave(x, repeats, axis)
-
-
-def tile(x, reps):
-    if isinstance(reps, _torch.Tensor):
-        reps = reps.detach().cpu().numpy().tolist()
-    return x.repeat(reps)
 
 
 # noinspection PyUnresolvedReferences
