+++ ivy/functional/ivy/old/general.py
@@ -219,30 +219,6 @@ def arange(stop: Number, start: Number = 0, step: Number = 1, dtype: ivy.Dtype =
     return _cur_framework().arange(stop, start, step, dtype, dev)
 
 
-# noinspection PyShadowingNames
-def linspace(start: Union[ivy.Array, ivy.NativeArray, Number], stop: Union[ivy.Array, ivy.NativeArray, Number],
-             num: int, axis: int = None, dev: ivy.Device = None)\
-        -> Union[ivy.Array, ivy.NativeArray]:
-    """
-    Generates a certain number of evenly-spaced values in an interval along a given axis.
-
-    See :math:`arange` that allows to specify the step size of evenly spaced values in an interval.
-
-    :param start: First entry in the range.
-    :type start: array
-    :param stop: Final entry in the range.
-    :type stop: array
-    :param num: Number of values to generate.
-    :type num: int
-    :param axis: Axis along which the operation is performed.
-    :type axis: int
-    :param dev: device on which to create the array 'cuda:0', 'cuda:1', 'cpu' etc.
-    :type dev: ivy.Device
-    :return: Tensor of evenly-spaced values.
-    """
-    return _cur_framework(start).linspace(start, stop, num, axis, dev)
-
-
 
 
 
@@ -280,55 +256,6 @@ def stack(xs: Iterable[Union[ivy.Array, ivy.NativeArray]], axis: int = 0)\
 
 
 
-def split(x: Union[ivy.Array, ivy.NativeArray], num_or_size_splits: Union[int, Iterable[int]] = None, axis: int = 0,
-          with_remainder: bool = False) -> Union[ivy.Array, ivy.NativeArray]:
-    """
-    Splits an array into multiple sub-arrays.
-
-    :param x: Tensor to be divided into sub-arrays.
-    :type x: array
-    :param num_or_size_splits: Number of equal arrays to divide the array into along the given axis if an integer.
-                               The size of each split element if a sequence of integers.
-                               Default is to divide into as many 1-dimensional arrays as the axis dimension.
-    :type num_or_size_splits: int, optional
-    :param axis: The axis along which to split, default is 0.
-    :type axis: int, optional
-    :param with_remainder: If the tensor does not split evenly, then store the last remainder entry. Default is False.
-    :type with_remainder: bool, optional
-    :return: A list of sub-arrays.
-    """
-    return _cur_framework(x).split(x, num_or_size_splits, axis, with_remainder)
-
-
-def repeat(x: Union[ivy.Array, ivy.NativeArray], repeats: Union[int, Iterable[int]], axis: int = None)\
-        -> Union[ivy.Array, ivy.NativeArray]:
-    """
-    Repeat values along a given dimension
-
-    :param x: Input array.
-    :type x: array
-    :param repeats: The number of repetitions for each element. repeats is broadcast to fit the shape of the given axis.
-    :type repeats: int or sequence of ints.
-    :param axis: The axis along which to repeat values.
-                  By default, use the flattened input array, and return a flat output array.
-    :type axis: int, optional
-    :return: The repeated output array.
-    """
-    return _cur_framework(x).repeat(x, repeats, axis)
-
-
-def tile(x: Union[ivy.Array, ivy.NativeArray], reps: Iterable[int])\
-        -> Union[ivy.Array, ivy.NativeArray]:
-    """
-    Constructs an array by repeating x the number of times given by reps.
-
-    :param x: Input array.
-    :type x: array
-    :param reps: The number of repetitions of x along each axis.
-    :type reps: sequence of ints
-    :return: The tiled output array.
-    """
-    return _cur_framework(x).tile(x, reps)
 
 
 def constant_pad(x: Union[ivy.Array, ivy.NativeArray], pad_width: Iterable[Tuple[int]], value: Number = 0)\
